requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
elasticdeform
pandas
scipy
scikit-image
